API Rename TopPage classes
