+++ _config/config.yml
@@ -26,7 +26,7 @@ SilverStripe\CMS\Controllers\CMSMain:
 
 SilverStripe\CMS\Model\SiteTree:
   extensions:
-    topPageSiteTreeExtension: DNADesign\Elemental\TopPage\SiteTreeExtension
+    topPageSiteTreeExtension: DNADesign\Elemental\Extensions\TopPageSiteTreeExtension
 
 SilverStripe\Versioned\VersionedGridFieldItemRequest:
   extensions:
