+++ lang/sl.yml
@@ -66,9 +66,9 @@ sl:
     Summary: Povzetek
     Title: Naslov
     Type: 'Tip gradnika'
-  DNADesign\Elemental\TopPage\DataExtension:
+  DNADesign\Elemental\Extensions\TopPageElementExtension:
     has_one_TopPage: 'Vrhnja stran'
-  DNADesign\Elemental\TopPage\FluentExtension:
+  DNADesign\Elemental\Extensions\TopPageFluentElementExtension
     db_TopPageLocale: 'Lokalizacija vrhnje strani '
   SilverStripe\CMS\Controllers\CMSPageHistoryController:
     PREVIEW: Predogled
