+++ src/Models/ElementalArea.php
@@ -4,7 +4,7 @@ namespace DNADesign\Elemental\Models;
 
 use DNADesign\Elemental\Controllers\ElementController;
 use DNADesign\Elemental\Extensions\ElementalAreasExtension;
-use DNADesign\Elemental\TopPage\DataExtension;
+use DNADesign\Elemental\Extensions\TopPageElementExtension;
 use SilverStripe\Core\ClassInfo;
 use SilverStripe\Core\Injector\Injector;
 use SilverStripe\Dev\TestOnly;
@@ -37,7 +37,7 @@ class ElementalArea extends DataObject
 
     private static $extensions = [
         Versioned::class,
-        DataExtension::class,
+        TopPageElementExtension::class,
     ];
 
     private static $owns = [
